requirements.txt
asgiref==3.8.1
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.5.1
